added support for authenticated v2 rest requests, fixed request URIs

+++ v2/rest/platform_status.go
@@ -11,12 +11,12 @@ func (p *PlatformService) Status() (bool, error) {
 	if err != nil {
 		return false, err
 	}
-/*
-	// raw is an interface type, but we only care about len & index 0
-	s := make([]int, len(raw))
-	for i, v := range raw {
-		s[i] = v.(int)
-	}
-*/
-	return len(raw) > 0 && raw[0].(int) == 1, nil
+	/*
+		// raw is an interface type, but we only care about len & index 0
+		s := make([]int, len(raw))
+		for i, v := range raw {
+			s[i] = v.(int)
+		}
+	*/
+	return len(raw) > 0 && raw[0].(float64) == 1, nil
 }
